Nov 21- After Lab Commmit
CHECK PPM. There is a misalignment when reading the Y values.
We have to properly check to make sure the values being read in are
given enough time to be used.
The column error still exists. It's skipping a few of the end pixels,
causing the the next line pixels to be used to the picture.